+++ eveuniverse/tests/test_models_3.py
@@ -2,9 +2,9 @@ from unittest.mock import patch
 
 import requests_mock
 
-from ..constants import EveGroupId
-from ..core import evesdeapi
-from ..models import (
+from eveuniverse.constants import EveGroupId
+from eveuniverse.core import evesdeapi
+from eveuniverse.models import (
     EveAsteroidBelt,
     EveIndustryActivityDuration,
     EveIndustryActivityMaterial,
@@ -19,7 +19,8 @@ from ..models import (
     EveType,
     EveTypeMaterial,
 )
-from ..utils import NoSocketsTestCase
+from eveuniverse.utils import NoSocketsTestCase
+
 from .testdata.esi import EsiClientStub
 from .testdata.sde import cache_content, sde_data, type_materials_cache_content
 
@@ -978,7 +979,7 @@ class TestEveSolarSystemBulkWithSection(NoSocketsTestCase):
                     enabled_sections=EveSolarSystem.enabled_sections.planets
                 ).values_list("id", flat=True)
             ),
-            {30000142, 30001161, 30045339, 30045342, 31000005},
+            {30000142, 30001161, 30045339, 30045342, 31000005, 30000157},
         )
         self.assertEqual(
             set(EvePlanet.objects.values_list("id", flat=True)),
@@ -1003,7 +1004,7 @@ class TestEveSolarSystemBulkWithSection(NoSocketsTestCase):
                     enabled_sections=EveSolarSystem.enabled_sections.planets
                 ).values_list("id", flat=True)
             ),
-            {30000142, 30001161, 30045339, 30045342, 31000005},
+            {30000142, 30001161, 30045339, 30045342, 31000005, 30000157},
         )
         self.assertEqual(
             set(EvePlanet.objects.values_list("id", flat=True)),